Add handler to handle auth request from frontend
Backend will receive a auth payload with username and password, then
check the user if is already existed in the boltdb and the password is
correct. If the user is authed, backend will return a jwt string to
the frontend. In the future, backend can fetch the user information from
the jwt string which come with the http request from frontend